добавил запрос на получение постов в отдельный компонент, поменял название типа для id favorites в resucer
