+++ src/store/postsReducer/action.ts
@@ -2,19 +2,12 @@ import {CardInfo} from "./type";
 
 export enum ACTION_POST_TYPES {
     ADD_POSTS = 'ADD_POSTS',
-    TOGGLE_FAVOURITE = "TOGGLE_FAVOURITE"
+    GET_ID_FAVOURITE = "GET_ID_FAVOURITE"
 }
 
 export const addPosts = (postsArray: CardInfo[]) => ({
     type: ACTION_POST_TYPES.ADD_POSTS,
     payload: {
         posts: postsArray
-    }
-})
-
-export const toggleFavorite = (id: number) => ({
-    type: ACTION_POST_TYPES.TOGGLE_FAVOURITE,
-    payload: {
-        id
     }
 })
