+++ src/store/postsReducer/postReducer.ts
@@ -10,8 +10,8 @@ export const postReducer = (state = initialPostState, action: ActionPostType) =>
     switch (action.type) {
         case ACTION_POST_TYPES.ADD_POSTS:
             return {...state, posts: action.payload.posts}
-        case ACTION_POST_TYPES.TOGGLE_FAVOURITE:
-            return ({...state, posts: action.payload.id})
+        case ACTION_POST_TYPES.GET_ID_FAVOURITE:
+            return {...state, posts: action.payload.id}
         default:
             return state
     }
